+++ src/main/kotlin/org/jetbrains/changelog/ChangelogPlugin.kt
@@ -3,6 +3,7 @@ package org.jetbrains.changelog
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.jetbrains.changelog.tasks.GetChangelogTask
+import org.jetbrains.changelog.tasks.InitializeChangelogTask
 import org.jetbrains.changelog.tasks.PatchChangelogTask
 
 class ChangelogPlugin : Plugin<Project> {
@@ -14,12 +15,18 @@ class ChangelogPlugin : Plugin<Project> {
 
         project.run {
             extensions.create("changelog", ChangelogPluginExtension::class.java, project)
-            tasks.create("patchChangelog", PatchChangelogTask::class.java) {
-                it.group = "changelog"
-            }
-            tasks.create("getChangelog", GetChangelogTask::class.java) {
-                it.group = "changelog"
-                it.outputs.upToDateWhen { false }
+
+            tasks.apply {
+                create("patchChangelog", PatchChangelogTask::class.java) {
+                    it.group = "changelog"
+                }
+                create("initializeChangelog", InitializeChangelogTask::class.java) {
+                    it.group = "changelog"
+                }
+                create("getChangelog", GetChangelogTask::class.java) {
+                    it.group = "changelog"
+                    it.outputs.upToDateWhen { false }
+                }
             }
         }
     }